Moved side panels to their own file
