+++ src/lib/components/sidePanel/SidePanelEnum.ts
@@ -1,0 +1,4 @@
+export enum SidePanelEnum {
+	Left,
+	Right,
+}
